+++ tests/scripts/script_all_errors.py
@@ -22,4 +22,6 @@ if __name__ == '__main__':
         color_code = weechat.info_get('irc_nick_color', 'nick')
         color_name = weechat.info_get('irc_nick_color_name', 'nick')
         weechat.hook_modifier('irc_in_privmsg', 'callback', '')
+        weechat.hook_signal('*,irc_out_privmsg', 'callback', '')
+        weechat.hook_signal('*,irc_outtags_privmsg', 'callback', '')
         sys.exit(1)