+++ tests/scripts/script_signal_irc_out.py
@@ -1,0 +1,11 @@
+#
+# Author: Sébastien Helleu <flashcode@flashtux.org>
+#
+
+"""A WeeChat script."""
+
+import weechat
+
+if __name__ == '__main__':
+    if weechat.register('script', 'author', '0.1', 'GPL3', 'desc', '', ''):
+        weechat.hook_signal('*,irc_out_privmsg', 'callback', '')