+++ tests/test_main.py
@@ -102,7 +102,7 @@ def test_main_dir():
         with mock.patch.object(sys, 'argv', args):
             weechat_script_lint.main()
     assert exc.type == SystemExit
-    assert exc.value.code == 19
+    assert exc.value.code == 23
 
     args = [
         'weechat-script-lint',