+++ ChangeLog.md
@@ -5,6 +5,7 @@
 - Sort messages by severity (high first).
 - Fix duplicate errors on `python2_bin`.
 - Check also empty files.
+- Add warnings on signals `irc_out_xxx` and `irc_outtags_xxx`.
 - Add warning on modifier `irc_in_xxx`.
 - Add warnings on deprecated info `irc_nick_color` and `irc_nick_color_name`.
 - Add warnings on deprecated functions `hook_completion_get_string` and `hook_completion_list_add`.